Move metrics into their own directory
While they are notionally a service, it seems cleaner to isolate them
this way--especially with the API breakout looming.
