+++ app/metrics/counters.rb
@@ -1,6 +1,6 @@
-require 'metrics/counter_support'
+require 'support/counter_support'
 
-module Metrics
+module Counters
   class CountVisits < ActiveRecord::Base
     def self.run
       first.count
@@ -8,22 +8,20 @@ module Metrics
   end
 
   class CountVisitsByState < ActiveRecord::Base
-    def self.run
-      all.each_with_object({}){ |counter, result|
-        result[counter.processing_state] = counter.count
-      }
+    def self.fetch_and_format
+      pluck(:processing_state, :count).to_h
     end
   end
 
   class CountVisitsByPrisonAndState < ActiveRecord::Base
-    extend Metrics::CounterSupport
+    extend CounterSupport
   end
 
   class CountVisitsByPrisonAndCalendarWeek < ActiveRecord::Base
-    extend Metrics::CounterSupport
+    extend CounterSupport
   end
 
   class CountVisitsByPrisonAndCalendarDate < ActiveRecord::Base
-    extend Metrics::CounterSupport
+    extend CounterSupport
   end
 end
